Should be fully runnable beginning to end...
Reads in all 15000 files from TDS_data_15000, then organizes the parameters so it can be fed into TSFresh's feature extraction tool. NaN values are dropped from detrap_en and def_conc, and the mean of each target data parameter is taken for each spectrum, so each spectrum corresponds to a single index in feature_select.

The TDS data values are too large and give a SizeError in feature_extract so the des_flux data is normalized with StandardScaler before featurizing. 

A copy of the data features, plus relevance tables for def_conc and detrap_en for TDS_data_15000 are in the Google Drive under NANO 120B/Data/TDS_data_featurized.

The last two cells are experimentation with applying Poisson's statistical noise onto the dataset. Next steps are playing with both Poisson and Gaussian noise parameters to fit Anze's expectations. Once that's done we can attempt to featurize the noisy data and train the model on that.