Added MLP with Poisson and Gauss Noises
I also wrote the initial for loop to replace the NaN with the average of the column before adding it to the dictionary and before concatenating into a data frame.  For the Gauss Noise MLP NN, I used grid search to find the hyperparameters before testing with the test data. This is goo to go. Only need to test it with the experimental data to be triple sure it works right.