Updated with Gauss noise implementation based...
on Anze's suggestion. The loop within the "Adding Gaussian Noise to des_flux" cell is the official structure going forward for any noise implementation.

The loop reads through the entire dataset, applying noise with distribution sigma_rand: a float between 1E14 and 1E17.
The subloop then applies noise with sigma_rand std to each datapoint, with the mean being the value of that datapoint. 

This applies noise with equal std to an entire spectrum, with each spectrum having a different level of noise. This best simulates the instrumental noise experienced when calibrating a TDS device. 

Next steps: Updating file readin and spectrum id to accommodate both the 0.5 K/s and 0.3 K/s datasets, then extract and select features. This will be the final fully preprocessed dataset.